code cleanup; finished refactor so that repositories use interfaces
