+++ cassandra/src/test/java/com/strategicgains/docussandra/handler/IndexCreatedHandlerTest.java
@@ -22,10 +22,10 @@ import com.strategicgains.docussandra.domain.Document;
 import com.strategicgains.docussandra.domain.Index;
 import com.strategicgains.docussandra.event.IndexCreatedEvent;
 import com.strategicgains.docussandra.domain.Table;
-import com.strategicgains.docussandra.persistence.DocumentRepository;
-import com.strategicgains.docussandra.persistence.IndexRepository;
-import com.strategicgains.docussandra.persistence.IndexStatusRepository;
-import com.strategicgains.docussandra.persistence.IndexStatusRepositoryTest;
+import com.strategicgains.docussandra.persistence.impl.DocumentRepositoryImpl;
+import com.strategicgains.docussandra.persistence.impl.IndexRepositoryImpl;
+import com.strategicgains.docussandra.persistence.impl.IndexStatusRepositoryImpl;
+import com.strategicgains.docussandra.persistence.impl.IndexStatusRepositoryImplTest;
 import com.strategicgains.docussandra.testhelper.Fixtures;
 import java.io.IOException;
 import java.util.Date;
@@ -48,12 +48,12 @@ import org.slf4j.LoggerFactory;
 public class IndexCreatedHandlerTest
 {
 
-    private static final Logger LOGGER = LoggerFactory.getLogger(IndexStatusRepositoryTest.class);
+    private static final Logger LOGGER = LoggerFactory.getLogger(IndexStatusRepositoryImplTest.class);
     private Fixtures f;
 
-    private IndexRepository indexRepo;
-    private IndexStatusRepository statusRepo;
-    private DocumentRepository docRepo;
+    private IndexRepositoryImpl indexRepo;
+    private IndexStatusRepositoryImpl statusRepo;
+    private DocumentRepositoryImpl docRepo;
 
     public IndexCreatedHandlerTest() throws Exception
     {
@@ -80,9 +80,9 @@ public class IndexCreatedHandlerTest
         f.insertDatabase(testDb);
         f.insertTable(Fixtures.createTestTable());
         f.insertDocuments(Fixtures.getBulkDocuments());
-        indexRepo = new IndexRepository(f.getSession());
-        statusRepo = new IndexStatusRepository(f.getSession());
-        docRepo = new DocumentRepository(f.getSession());
+        indexRepo = new IndexRepositoryImpl(f.getSession());
+        statusRepo = new IndexStatusRepositoryImpl(f.getSession());
+        docRepo = new DocumentRepositoryImpl(f.getSession());
     }
 
     @After
