+++ integration_test/src/test/java/com/strategicgains/docussandra/controller/DatabaseControllerTest.java
@@ -19,11 +19,11 @@ import com.jayway.restassured.RestAssured;
 import static com.jayway.restassured.RestAssured.expect;
 import static com.jayway.restassured.RestAssured.given;
 import com.strategicgains.docussandra.domain.Database;
-import com.strategicgains.docussandra.persistence.DocumentRepository;
+import com.strategicgains.docussandra.persistence.impl.DocumentRepositoryImpl;
 import com.strategicgains.docussandra.persistence.ITableRepository;
 import com.strategicgains.docussandra.persistence.impl.ITableRepositoryImpl;
-import com.strategicgains.docussandra.persistence.IndexRepository;
-import com.strategicgains.docussandra.persistence.TableRepository;
+import com.strategicgains.docussandra.persistence.impl.IndexRepositoryImpl;
+import com.strategicgains.docussandra.persistence.impl.TableRepositoryImpl;
 import org.junit.BeforeClass;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -215,16 +215,16 @@ public class DatabaseControllerTest
         expect().statusCode(404).when()
                 .get(testDb.name());
         //check table deletion (using direct db calls instead of REST-- being slightly lazy here)
-        TableRepository tableRepo = new TableRepository(f.getSession());
+        TableRepositoryImpl tableRepo = new TableRepositoryImpl(f.getSession());
         assertFalse(tableRepo.exists(Fixtures.createTestTable().getId()));
         //check index deletion (using direct db calls instead of REST-- being slightly lazy here)
-        IndexRepository indexRepo = new IndexRepository(f.getSession());
+        IndexRepositoryImpl indexRepo = new IndexRepositoryImpl(f.getSession());
         assertFalse(indexRepo.exists(Fixtures.createTestIndexOneField().getId()));
         //check iTable deletion (using direct db calls instead of REST-- being slightly lazy here)
         ITableRepository iTableRepo = new ITableRepositoryImpl(f.getSession());
         assertFalse(iTableRepo.iTableExists(Fixtures.createTestIndexOneField()));
         //check document deletion (using direct db calls instead of REST-- being slightly lazy here)
-        DocumentRepository docRepo = new DocumentRepository(f.getSession());
+        DocumentRepositoryImpl docRepo = new DocumentRepositoryImpl(f.getSession());
         boolean expectedExceptionThrown = false;
         try
         {
