+++ rest/src/main/java/com/strategicgains/docussandra/config/Configuration.java
@@ -24,8 +24,14 @@ import com.strategicgains.docussandra.persistence.DatabaseRepository;
 import com.strategicgains.docussandra.persistence.DocumentRepository;
 import com.strategicgains.docussandra.persistence.IndexRepository;
 import com.strategicgains.docussandra.persistence.IndexStatusRepository;
-import com.strategicgains.docussandra.persistence.impl.QueryRepositoryImpl;
+import com.strategicgains.docussandra.persistence.QueryRepository;
 import com.strategicgains.docussandra.persistence.TableRepository;
+import com.strategicgains.docussandra.persistence.impl.DatabaseRepositoryImpl;
+import com.strategicgains.docussandra.persistence.impl.DocumentRepositoryImpl;
+import com.strategicgains.docussandra.persistence.impl.IndexRepositoryImpl;
+import com.strategicgains.docussandra.persistence.impl.IndexStatusRepositoryImpl;
+import com.strategicgains.docussandra.persistence.impl.QueryRepositoryImpl;
+import com.strategicgains.docussandra.persistence.impl.TableRepositoryImpl;
 import com.strategicgains.docussandra.service.DatabaseService;
 import com.strategicgains.docussandra.service.DocumentService;
 import com.strategicgains.docussandra.service.IndexService;
@@ -104,12 +110,12 @@ public class Configuration
         {
             LOGGER.error("Could not init database; trying to continue startup anyway (in case DB was manually created).", e);
         }
-        DatabaseRepository databaseRepository = new DatabaseRepository(dbConfig.getSession());
-        TableRepository tableRepository = new TableRepository(dbConfig.getSession());
-        DocumentRepository documentRepository = new DocumentRepository(dbConfig.getSession());
-        IndexRepository indexRepository = new IndexRepository(dbConfig.getSession());
-        QueryRepositoryImpl queryRepository = new QueryRepositoryImpl(dbConfig.getSession());
-        IndexStatusRepository indexStatusRepository = new IndexStatusRepository(dbConfig.getSession());
+        DatabaseRepository databaseRepository = new DatabaseRepositoryImpl(dbConfig.getSession());
+        TableRepository tableRepository = new TableRepositoryImpl(dbConfig.getSession());
+        DocumentRepository documentRepository = new DocumentRepositoryImpl(dbConfig.getSession());
+        IndexRepository indexRepository = new IndexRepositoryImpl(dbConfig.getSession());
+        QueryRepository queryRepository = new QueryRepositoryImpl(dbConfig.getSession());
+        IndexStatusRepository indexStatusRepository = new IndexStatusRepositoryImpl(dbConfig.getSession());
 
         DatabaseService databaseService = new DatabaseService(databaseRepository);
         TableService tableService = new TableService(databaseRepository, tableRepository);
